// node-projects/backups/nub/20250426/interface/types.ts
export type TPlanData = {
  providerId: ProviderId;
  planCode: string;
  amount: number;
};
// ...
export enum ProviderId {
  ProviderX = "providerX", // TODO: remove; test provider
}